Add styles for header and hero content on page
The top logo and links are place in right right
position through flexbox. The "hero content"
has gotten couple of styles but needs more tinker-
ing.

Still trying to get all the parts in the right
order, there is a lot to fiddle around with.
But I think I'm getting there.